added pdfs to decribe the theory behind the algorithms
Continuum morphometry describes a formalisms that extendsa kind of quasiconformal transformations to a setting of surfaces in embeddede in 3D. The PDF describes both the problem of matching two nearly identical surface, which is relavant to growing surfaces. And the problem of registering different surfaces. The distance given in this case includes elastic distortions in the process of matching in addition to gradients in those distortions, which has not been considered before.

The second PDF talks about the embedding of a 2D elastic mesh given the metric you want to embed. For example, if you compute your metric using a square root velocity function as described in the book "Functional and Shape Data Analysis", this code can embedd that metric optimally in 2D.