curve morphometry and eggs
"Curve_Morphometry.py" computes the registration between different curves (given as the curvature and torsion as function of arc length). This also uses a gradient cost in the energy which is relatively novel.

The accompanying examples of egg shape parameters appeared in a science publication on egg shapes titled: "Avian egg shape: Form, function, and evolution".

The algorithm analyses a set of shapes an returns a phylogenetic tree between them.